+++ pyproject.toml
@@ -1,5 +1,5 @@
 [tool.poetry]
-name = "spacereps"
+name = "server"
 version = "0.1.0"
 description = ""
 authors = ["gladnicks <samgladnick@hotmail.com>"]
